requirements.txt
Flask==2.3.2
pillow==10.3.0
pycairo==1.28.0
cairocffi==1.7.1
python-dotenv==1.0.1
google-generativeai==0.7.2
google-ai-generativelanguage==0.6.6
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
httpx==0.28.1
tenacity==8.3.0
google==3.0.0
google-auth==2.40.3
google-genai==1.36.0
openai==1.107.3

python-telegram-bot==22.4

